requirements.txt
streamlit
PyMuPDF
python-docx
google-generativeai
fastapi
uvicorn
python-multipart
